display cost, delete btn load/save, css

+++ app/app.js
@@ -8,6 +8,8 @@ const itemMultiplyDeux = document.querySelector('.multiple-deux-buy');
 const btnLoad = document.querySelector(".load");
 const btnSave = document.querySelector(".save");
 const btnReset = document.querySelector(".reset");
+const recap = document.querySelector("#clics-recap");
+
 
 /* BTN CLICKER */
 
@@ -22,13 +24,13 @@ if(localStorage.getItem("score") == null){
     scoreDisplay.innerHTML = score;
 }
 
-
 //variables click item
 let clickCost = 15;
 let click = 1;
 
-// variable auto item
+//variable auto item
 let autoCost = 100;
+let autoClick = 0;
 
 //variable multiply item
 let multiplyCost = 200;
@@ -37,7 +39,6 @@ let multiplyClick = 0;
 //variable multiplayDeux item
 let multiplyDeuxCost = 400;
 let multiplyDeuxClick = 0;
-
 
 /* FUNCTIONS */
 
@@ -48,6 +49,16 @@ let multiplyDeuxClick = 0;
 function addToScore(amount){
     score = score + amount;
     scoreDisplay.innerHTML = score;
+
+}
+
+/* La fonction recapBonus est appelée à chaque fois qu'un achat est fait
+   Cela permet d'incrémenter le bonus à chaque achat.
+ */
+let recapClicBonus;
+function recapBonus(){
+    let recapClicBonus = click + autoClick + multiplyClick + multiplyDeuxClick;
+    recap.innerHTML = recapClicBonus;
 }
 
 function save(){
@@ -56,10 +67,12 @@ function save(){
         click: click,
         clickCost: clickCost,
         autoCost: autoCost,
+        autoClick: autoClick,
         multiplyClick: multiplyClick,
         multiplyCost: multiplyCost,
         multiplyDeuxClick: multiplyDeuxClick,
         multiplyDeuxCost: multiplyDeuxCost,
+        recapClicBonus: recapClicBonus,
     };
     localStorage.setItem("gameSave", JSON.stringify(gameSave));
 }
@@ -67,7 +80,8 @@ function save(){
 setInterval(function(){
     save();
    scoreDisplay.innerHTML = score;
-}, 10000);
+}, 1000);
+
 
 /* Fonction qui va permettre de sauvegarder la partie
    Stockage dans la variable saveGame le getItem "gameSave"
@@ -80,17 +94,30 @@ function load(){
     if (typeof saveGame.click !== "undefined") click = saveGame.click;
     if (typeof saveGame.clickCost !== "undefined") clickCost = saveGame.clickCost;
     if (typeof saveGame.autoCost !== "undefined") autoCost = saveGame.autoCost;
+    if (typeof saveGame.autoClick !== "undefined") autoClick = saveGame.autoClick;
     if (typeof saveGame.multiplyClick !== "undefined") multiplyClick = saveGame.multiplyClick;
     if (typeof saveGame.multiplyCost !== "undefined") multiplyCost = saveGame.multiplyCost;
     if (typeof saveGame.multiplyDeuxClick !== "undefined") multiplyDeuxClick = saveGame.multiplyDeuxClick;
     if (typeof saveGame.multiplyDeuxCost !== "undefined") multiplyDeuxCost = saveGame.multiplyDeuxCost;
+    if (typeof saveGame.recapClicBonus !== "undefined") recapClicBonus = saveGame.recapClicBonus;
     scoreDisplay.innerHTML = score
+    recap.innerHTML = saveGame.recapClicBonus;
 }
 window.onload = function(){
     load();
+    recapBonus()
+    document.getElementById('click').innerHTML = click;
+    document.getElementById('clickCost').innerHTML = clickCost;
+    document.getElementById('autoCost').innerHTML = autoCost;
+    document.getElementById('autoClick').innerHTML = autoClick + 1;
+    document.getElementById('multiplyClick').innerHTML = multiplyClick + 2;
+    document.getElementById('multiplyCost').innerHTML = multiplyCost
+    document.getElementById('multiplyDeuxClick').innerHTML = multiplyDeuxClick + 4;
+    document.getElementById('multiplyDeuxCost').innerHTML = multiplyDeuxCost
 }
 
 function reset(){
+    if(confirm('Etes-vous certain de vouloir reset le jeu ?'))
     localStorage.clear();
     window.location = 'http://localhost:8888/clicker-game/clicker.php';
 
@@ -113,12 +140,19 @@ function buyClick(){
         scoreDisplay.innerHTML = score;
         document.getElementById('click').innerHTML = click;
         document.getElementById('clickCost').innerHTML = clickCost;
+        recapBonus();
     }else{
         alert("Vous n'avez pas assez de clics");
     }
 }
 function buyAutoClick(){
     if(score >= autoCost ){
+        score = score - autoCost;
+        autoClick = autoClick + 1
+        autoCost = Math.round(autoCost * 1.4);
+        document.getElementById('autoCost').innerHTML = autoCost;
+        document.getElementById('autoClick').innerHTML = autoClick + 1;
+        recapBonus();
         setInterval(() => {
             score = score + 1;
             document.querySelector('#score').innerHTML = score;
@@ -131,11 +165,13 @@ function buyAutoClick(){
 function buyMultiply(){
     if(score >= multiplyCost){
         score = score - multiplyCost;
-        multiplyClick = multiplyClick + 1;
+        multiplyClick = multiplyClick + 2;
         multiplyCost = Math.round(multiplyCost * 1.4);
         scoreDisplay.innerHTML = score;
-        document.getElementById('multiplyClick').innerHTML = multiplyClick;
+        document.getElementById('multiplyClick').innerHTML = multiplyClick + 2;
+        console.log(document.getElementById('multiplyClick'))
         document.getElementById('multiplyCost').innerHTML = multiplyCost
+        recapBonus();
     }else{
         alert("Vous n'avez pas assez de smiles pour faire cet achat !")
     }
@@ -147,8 +183,10 @@ function buyMultiplyDeux(){
         multiplyDeuxClick = multiplyDeuxClick + 4;
         multiplyDeuxCost = Math.round(multiplyDeuxCost * 1.7);
         scoreDisplay.innerHTML = score;
-        document.getElementById('multiplyDeuxClick').innerHTML = multiplyDeuxClick;
+        document.getElementById('multiplyDeuxClick').innerHTML = multiplyDeuxClick + 4;
+        console.log(document.getElementById('multiplyDeuxClick'))
         document.getElementById('multiplyDeuxCost').innerHTML = multiplyDeuxCost
+        recapBonus();
     }else{
         alert("Vous n'avez pas assez de smiles pour faire cet achat !")
     }
@@ -161,7 +199,7 @@ function buyMultiplyDeux(){
    montants à rajouter au score
  */
 btnClicker.addEventListener('click', () => {
-    addToScore(click + multiplyClick + multiplyDeuxClick);
+    addToScore(click + multiplyClick + multiplyDeuxClick + autoClick);
 })
 
 itemBonus.addEventListener('click', () => {
@@ -178,14 +216,6 @@ itemMultiply.addEventListener('click', () => {
 
 itemMultiplyDeux.addEventListener('click', () => {
     buyMultiplyDeux();
-})
-
-btnLoad.addEventListener('click', () => {
-    load();
-})
-
-btnSave.addEventListener('click', () => {
-    save();
 })
 
 btnReset.addEventListener('click', () => {
